+++ lib/client_side_validations/active_model/format.rb
@@ -8,12 +8,12 @@ module ClientSideValidations
         if options[:with].respond_to?(:call)
           return unless force
 
-          options[:with] = options[:with].call(model)
+          options[:with] = resolve_proc(options[:with], model)
           build_client_side_hash(model, attribute, options)
         elsif options[:without].respond_to?(:call)
           return unless force
 
-          options[:without] = options[:without].call(model)
+          options[:without] = resolve_proc(options[:without], model)
           build_client_side_hash(model, attribute, options)
         else
           super